--- config.sample.js
@@ -1,6 +1,0 @@
-var config = {
-  telegramToken: 'MY_SUPER_SECRET_TOKEN',
-  welcomeMsg: 'Welcome to the jungle #{name}'
-};
-
-module.exports = config;
