+++ config/config.sample.js
@@ -1,0 +1,5 @@
+const config = {
+  telegramToken: 'MY_SUPER_SECRET_TOKEN'
+};
+
+module.exports = config;
