+++ config/messages.js
@@ -1,0 +1,8 @@
+var messages = {
+  welcomeMsg: 'Hola #{name}, bienvenido a ngVenezuela, te invitamos a ' +
+    'seguirnos en twitter `@ngVenezuela` y a unirte a nuestro grupo ' +
+    'en Google bit.ly/ng-venezuela-google-groups Puedes colocar tus ' +
+    'preguntas ahí, también tenemos cursos, de typescript e iniciacion con Angular.'
+};
+
+module.exports = messages;
