+++ package.json
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/ngVenezuela/wengy-ven#readme",
   "dependencies": {
-    "node-telegram-bot-api": "0.23.3"
+    "node-telegram-bot-api": "0.23.3",
+    "winston": "2.2.0"
   }
 }
